refactor(struct-init-v2): collect per-function field effects through a receiver
The per-function collection passes were free functions threading shared
state (boundary indices, classified locals, output sets) through up to
eight parameters each. Introduce functionCollector to own that state:
the passes become argument-light methods reading from the receiver, and
collectFunction reduces to constructing the collector and running its
passes. Also consolidate the duplicated IndexedFieldPath orderings onto
one compare method used by every sorted consumer.

No behavior change: fixtures and the exported fact encoding are
byte-identical.
