+++ assertion/function/structfieldeffects/effects.go
@@ -15,15 +15,13 @@
 package structfieldeffects
 
 import (
-	"go/ast"
-	"go/token"
+	"cmp"
 	"go/types"
+	"slices"
 	"sort"
 	"strings"
 
 	"go.uber.org/nilaway/annotation"
-	"go.uber.org/nilaway/util/analysishelper"
-	"go.uber.org/nilaway/util/asthelper"
 	"go.uber.org/nilaway/util/typeshelper"
 )
 
@@ -96,134 +94,6 @@ func fieldPathsForIndex(effects fieldEffects, funcObj *types.Func, idx int) []st
 	return paths
 }
 
-// collectedFieldEffects owns the unclosed package summary and the forwarding state needed to close
-// it after imported effects have been seeded.
-type collectedFieldEffects struct {
-	summary               *BoundaryFieldEffects
-	paramForwardingEdges  map[*types.Func][]paramFieldForwardEdge
-	returnForwardingEdges map[*types.Func][]returnForwardEdge
-	callees               map[*types.Func]bool
-}
-
-func newCollectedFieldEffects() *collectedFieldEffects {
-	return &collectedFieldEffects{
-		summary: &BoundaryFieldEffects{
-			ParamReads:    make(fieldEffects),
-			ReturnReads:   make(fieldEffects),
-			ParamWrites:   make(fieldEffects),
-			ReturnEffects: make(fieldEffects),
-		},
-		paramForwardingEdges:  make(map[*types.Func][]paramFieldForwardEdge),
-		returnForwardingEdges: make(map[*types.Func][]returnForwardEdge),
-		callees:               make(map[*types.Func]bool),
-	}
-}
-
-// computeBoundaryFieldEffects collects the unclosed boundary summary and forwarding state for every
-// function and method in the package.
-//
-// Reads are gathered from selector bases — to evaluate `base.Sel`, base must be non-nil, so the
-// field path of base is a read of whatever boundary value it roots at (a parameter → ParamReads,
-// or a struct-returning-call result local → ReturnReads). ast.Inspect visits nested selectors, so
-// every prefix of a deep access is recorded. Writes are gathered from assignment LHS field chains
-// rooted at a parameter or receiver. Every static call also records an arg→param forwarding
-// edge (which caller parameter, possibly at a nested field prefix, is passed as which callee
-// parameter). Concrete struct returns contribute nil field paths, while direct returns of
-// same-package call results contribute return-forwarding edges. Static callees are retained so the
-// analyzer imports only facts needed by calls in this package before closing the collection.
-//
-// Unresolvable (interface/func-value) callees are treated as mutating/dereferencing nothing
-// (under-report only).
-func computeBoundaryFieldEffects(pass *analysishelper.EnhancedPass) *collectedFieldEffects {
-	collected := newCollectedFieldEffects()
-	for _, file := range pass.Files {
-		for _, decl := range file.Decls {
-			fd, ok := decl.(*ast.FuncDecl)
-			if !ok || fd.Body == nil {
-				continue
-			}
-			collected.collectFunction(pass, fd)
-		}
-	}
-	return collected
-}
-
-func (c *collectedFieldEffects) collectFunction(pass *analysishelper.EnhancedPass, fd *ast.FuncDecl) {
-	funcObj, ok := pass.TypesInfo.ObjectOf(fd.Name).(*types.Func)
-	if !ok {
-		return
-	}
-	sig, ok := funcObj.Type().(*types.Signature)
-	if !ok {
-		return
-	}
-
-	paramIdx := make(map[*types.Var]int)
-	if recv := sig.Recv(); recv != nil {
-		paramIdx[recv] = annotation.ReceiverParamIndex
-	}
-	for i := range sig.Params().Len() {
-		paramIdx[sig.Params().At(i)] = i
-	}
-
-	// Locals bound directly to a struct-returning call, so a later dereference of the local's fields
-	// can be attributed to that callee's result (the return-read demand).
-	resultVars := collectStructResultVars(pass, fd.Body)
-	c.collectReturnEffects(pass, fd, funcObj, resultVars)
-
-	// Selectors assigned to (`a.y = ...`) write their selected field rather than read it, so they
-	// must not contribute value demand. ast.Inspect visits an AssignStmt before descending into its
-	// LHS, so the set is always populated before the selector itself is visited.
-	assignedSelectors := make(map[*ast.SelectorExpr]bool)
-	ast.Inspect(fd.Body, func(n ast.Node) bool {
-		switch n := n.(type) {
-		case *ast.AssignStmt:
-			for _, lhs := range n.Lhs {
-				if sel, ok := ast.Unparen(lhs).(*ast.SelectorExpr); ok {
-					assignedSelectors[sel] = true
-				}
-			}
-			collectParamFieldWrites(pass, n, paramIdx, funcObj, c.summary.ParamWrites)
-		case *ast.CallExpr:
-			if callee := collectParamForwardEdges(pass, n, paramIdx, funcObj, c.paramForwardingEdges); callee != nil {
-				c.callees[callee] = true
-			}
-		case *ast.SelectorExpr:
-			collectFieldReadDemand(
-				pass, n, paramIdx, resultVars, funcObj,
-				c.summary.ParamReads, c.summary.ReturnReads, assignedSelectors[n],
-			)
-		}
-		return true
-	})
-}
-
-// collectReturnEffects records this function's concrete nil result fields and return
-// forwarding edges. resultVars maps locals bound directly to struct-returning calls and is shared
-// with the later read-demand pass.
-func (c *collectedFieldEffects) collectReturnEffects(pass *analysishelper.EnhancedPass, fd *ast.FuncDecl, funcObj *types.Func, resultVars map[*types.Var]structResultSource) {
-	sig := funcObj.Signature()
-	// Fast path: no struct-shaped result can ever produce a concrete return effect or forwarding
-	// edge, so the body walk below would do nothing. Use the same predicate as
-	// collectConcreteReturnEffects to avoid a narrower type check.
-	hasStructResult := false
-	for result := range sig.Results().Variables() {
-		if typeshelper.AsDeeplyStruct(result.Type()) != nil {
-			hasStructResult = true
-			break
-		}
-	}
-	if !hasStructResult {
-		return
-	}
-	allocationVars := collectStructAllocationVars(pass, fd.Body)
-	stableVars := collectStableStructVars(pass, fd.Body, allocationVars, resultVars)
-	collectConcreteReturnEffects(
-		pass, fd.Body, funcObj, allocationVars, resultVars, stableVars,
-		c.summary.ReturnEffects, c.returnForwardingEdges,
-	)
-}
-
 func (c *collectedFieldEffects) close() *BoundaryFieldEffects {
 	closeParamFieldSets(c.summary.ParamWrites, c.paramForwardingEdges)
 	closeParamFieldSets(c.summary.ParamReads, c.paramForwardingEdges)
@@ -252,12 +122,7 @@ func (e fieldEffects) sortedPaths(funcObj *types.Func) []IndexedFieldPath {
 	for key := range e[funcObj] {
 		paths = append(paths, key)
 	}
-	sort.Slice(paths, func(i, j int) bool {
-		if paths[i].Idx != paths[j].Idx {
-			return paths[i].Idx < paths[j].Idx
-		}
-		return paths[i].Path < paths[j].Path
-	})
+	slices.SortFunc(paths, IndexedFieldPath.compare)
 	return paths
 }
 
@@ -267,6 +132,16 @@ func (e fieldEffects) sortedPaths(funcObj *types.Func) []IndexedFieldPath {
 type IndexedFieldPath struct {
 	Idx  int
 	Path string
+}
+
+// compare orders IndexedFieldPaths by index, then path. The boundary summaries are map-backed
+// sets that reach an exported fact, so every consumer sorts with this ordering to keep map
+// iteration order from leaking.
+func (p IndexedFieldPath) compare(other IndexedFieldPath) int {
+	return cmp.Or(
+		cmp.Compare(p.Idx, other.Idx),
+		cmp.Compare(p.Path, other.Path),
+	)
 }
 
 // fieldEffects maps each function to a set of boundary field paths.
@@ -283,439 +158,6 @@ func (e fieldEffects) add(funcObj *types.Func, key IndexedFieldPath) bool {
 	}
 	e[funcObj][key] = true
 	return true
-}
-
-// structResultSource identifies the struct-returning callee and result index a local variable was
-// assigned from, so dereferences of that local's fields can be attributed to the callee's return.
-type structResultSource struct {
-	callee typeshelper.StaticCallTarget
-	idx    int
-}
-
-// collectStructResultVars maps each local variable bound directly from a static struct-returning call
-// (`b := callee()`, `var b = callee()`, `b, err := callee()`) to the callee and result index it
-// came from. Only struct (or pointer-to-struct) results are recorded. Used to attribute return-read
-// demand to callees. A variable later reassigned is best-effort (first binding wins); a non-bare or
-// cross-package callee is skipped (under-report).
-func collectStructResultVars(pass *analysishelper.EnhancedPass, body *ast.BlockStmt) map[*types.Var]structResultSource {
-	out := make(map[*types.Var]structResultSource)
-	record := func(lhs, rhs []ast.Expr) {
-		if len(rhs) != 1 {
-			return
-		}
-		call, ok := ast.Unparen(rhs[0]).(*ast.CallExpr)
-		if !ok {
-			return
-		}
-		target, ok := typeshelper.ResolveStaticCallTarget(pass.TypesInfo, call)
-		if !ok {
-			return
-		}
-		if target.Signature.Results().Len() != len(lhs) {
-			return
-		}
-		for i, l := range lhs {
-			ident, ok := l.(*ast.Ident)
-			if !ok {
-				continue
-			}
-			v, ok := pass.TypesInfo.ObjectOf(ident).(*types.Var)
-			if !ok {
-				continue
-			}
-			if typeshelper.AsDeeplyStruct(target.Signature.Results().At(i).Type()) == nil {
-				continue
-			}
-			if _, seen := out[v]; !seen {
-				out[v] = structResultSource{callee: target, idx: i}
-			}
-		}
-	}
-	ast.Inspect(body, func(n ast.Node) bool {
-		switch n := n.(type) {
-		case *ast.AssignStmt:
-			record(n.Lhs, n.Rhs)
-		case *ast.ValueSpec:
-			// `var b = callee()` / `var b1, b2 = callee()`: the names are the LHS.
-			idents := make([]ast.Expr, len(n.Names))
-			for i, nm := range n.Names {
-				idents[i] = nm
-			}
-			record(idents, n.Values)
-		}
-		return true
-	})
-	return out
-}
-
-type structAllocationSource struct {
-	structType *types.Struct
-	fieldInits []ast.Expr
-}
-
-// staticStructAllocation recognizes struct values created directly by a composite literal or new.
-func staticStructAllocation(pass *analysishelper.EnhancedPass, expr ast.Expr) (structAllocationSource, bool) {
-	switch e := ast.Unparen(expr).(type) {
-	case *ast.UnaryExpr:
-		if e.Op == token.AND {
-			return staticStructAllocation(pass, e.X)
-		}
-	case *ast.CompositeLit:
-		if structType := typeshelper.AsDeeplyStruct(pass.TypesInfo.TypeOf(e)); structType != nil {
-			return structAllocationSource{structType: structType, fieldInits: e.Elts}, true
-		}
-	case *ast.CallExpr:
-		ident, ok := ast.Unparen(e.Fun).(*ast.Ident)
-		if ok && pass.TypesInfo.ObjectOf(ident) == typeshelper.BuiltinNew {
-			if structType := typeshelper.AsDeeplyStruct(pass.TypesInfo.TypeOf(e)); structType != nil {
-				return structAllocationSource{structType: structType}, true
-			}
-		}
-	}
-	return structAllocationSource{}, false
-}
-
-// collectStructAllocationVars records locals whose defining value is a concrete struct allocation.
-func collectStructAllocationVars(pass *analysishelper.EnhancedPass, body *ast.BlockStmt) map[*types.Var]structAllocationSource {
-	var out map[*types.Var]structAllocationSource
-	recordAllocation := func(ident *ast.Ident, source structAllocationSource) {
-		v, ok := pass.TypesInfo.Defs[ident].(*types.Var)
-		if !ok {
-			return
-		}
-		if out == nil {
-			out = make(map[*types.Var]structAllocationSource)
-		}
-		out[v] = source
-	}
-	recordIfAllocation := func(ident *ast.Ident, expr ast.Expr) {
-		source, ok := staticStructAllocation(pass, expr)
-		if ok {
-			recordAllocation(ident, source)
-		}
-	}
-	ast.Inspect(body, func(n ast.Node) bool {
-		switch n := n.(type) {
-		case *ast.AssignStmt:
-			if len(n.Lhs) != len(n.Rhs) {
-				return true
-			}
-			for i, lhs := range n.Lhs {
-				if ident, ok := ast.Unparen(lhs).(*ast.Ident); ok {
-					recordIfAllocation(ident, n.Rhs[i])
-				}
-			}
-		case *ast.ValueSpec:
-			if len(n.Values) == 0 {
-				for _, ident := range n.Names {
-					// `var x *A` declares a nil pointer, not a zero-valued struct: it has no
-					// field shape to summarize
-					if typeshelper.IsDeeplyType[*types.Pointer](pass.TypesInfo.TypeOf(ident)) {
-						continue
-					}
-					if structType := typeshelper.AsDeeplyStruct(pass.TypesInfo.TypeOf(ident)); structType != nil {
-						recordAllocation(ident, structAllocationSource{structType: structType})
-					}
-				}
-				return true
-			}
-			if len(n.Names) != len(n.Values) {
-				return true
-			}
-			for i, ident := range n.Names {
-				recordIfAllocation(ident, n.Values[i])
-			}
-		case *ast.FuncLit:
-			return false
-		}
-		return true
-	})
-	return out
-}
-
-// collectStableStructVars returns candidate locals whose only uses are as bare operands of explicit
-// returns from the enclosing function. Starting from known-safe uses makes unfamiliar syntax
-// conservative by default: any reassignment, mutation, alias, observation, escape, or closure
-// capture is a non-return use and disqualifies the local.
-func collectStableStructVars(pass *analysishelper.EnhancedPass, body *ast.BlockStmt, allocationVars map[*types.Var]structAllocationSource, resultVars map[*types.Var]structResultSource) map[*types.Var]bool {
-	if len(allocationVars) == 0 && len(resultVars) == 0 {
-		return nil
-	}
-
-	candidates := make(map[*types.Var]bool, len(allocationVars)+len(resultVars))
-	for v := range allocationVars {
-		candidates[v] = true
-	}
-	for v := range resultVars {
-		candidates[v] = true
-	}
-
-	// Record exact identifier occurrences that are bare return operands. Returns inside closures do
-	// not return from the function being summarized, so they are not allowed uses.
-	allowedUses := make(map[*ast.Ident]bool)
-	ast.Inspect(body, func(n ast.Node) bool {
-		switch n := n.(type) {
-		case *ast.FuncLit:
-			return false
-		case *ast.ReturnStmt:
-			for _, result := range n.Results {
-				if ident, ok := ast.Unparen(result).(*ast.Ident); ok {
-					allowedUses[ident] = true
-				}
-			}
-		}
-		return true
-	})
-
-	// Only candidates that are actually returned start stable. The second walk intentionally enters
-	// closures so a captured candidate is seen as a disallowed use.
-	stable := make(map[*types.Var]bool)
-	for ident := range allowedUses {
-		if v, ok := pass.TypesInfo.Uses[ident].(*types.Var); ok && candidates[v] {
-			stable[v] = true
-		}
-	}
-
-	// Reject a candidate if any of its uses is not one of the exact return occurrences recorded
-	// above. Definitions are absent from TypesInfo.Uses, so they do not disqualify the candidate.
-	ast.Inspect(body, func(n ast.Node) bool {
-		ident, ok := n.(*ast.Ident)
-		if !ok {
-			return true
-		}
-		v, ok := pass.TypesInfo.Uses[ident].(*types.Var)
-		if ok && candidates[v] && !allowedUses[ident] {
-			delete(stable, v)
-		}
-		return true
-	})
-	return stable
-}
-
-// collectConcreteReturnEffects records concrete nil result fields and forwarding edges.
-func collectConcreteReturnEffects(
-	pass *analysishelper.EnhancedPass,
-	body *ast.BlockStmt,
-	funcObj *types.Func,
-	allocationVars map[*types.Var]structAllocationSource,
-	resultVars map[*types.Var]structResultSource,
-	stableVars map[*types.Var]bool,
-	effects fieldEffects,
-	edges map[*types.Func][]returnForwardEdge,
-) {
-	sig := funcObj.Signature()
-	addEdge := func(callerResultIdx int, target typeshelper.StaticCallTarget, calleeResultIdx int) {
-		if target.Origin == nil {
-			return
-		}
-		if calleeResultIdx < 0 || calleeResultIdx >= target.Signature.Results().Len() ||
-			typeshelper.AsDeeplyStruct(target.Signature.Results().At(calleeResultIdx).Type()) == nil {
-			return
-		}
-		edges[funcObj] = append(edges[funcObj], returnForwardEdge{
-			callerResultIdx: callerResultIdx,
-			callee:          target.Origin,
-			calleeResultIdx: calleeResultIdx,
-		})
-	}
-
-	ast.Inspect(body, func(n ast.Node) bool {
-		switch n := n.(type) {
-		case *ast.FuncLit:
-			return false
-		case *ast.ReturnStmt:
-			// A bare spreading return such as `return f()` is one expression that yields every
-			// result, so len(n.Results) is 1 while the signature may have many; we bail out
-			// rather than try to split it, so such multi-result call returns are unsupported.
-			if len(n.Results) != sig.Results().Len() {
-				return false
-			}
-			for resultIdx, resultExpr := range n.Results {
-				if typeshelper.AsDeeplyStruct(sig.Results().At(resultIdx).Type()) == nil {
-					continue
-				}
-				if source, ok := staticStructAllocation(pass, resultExpr); ok {
-					enumerateConcreteReturnEffects(pass, funcObj, resultIdx, source.structType, source.fieldInits, "", effects)
-					continue
-				}
-				if call, ok := ast.Unparen(resultExpr).(*ast.CallExpr); ok {
-					target, ok := typeshelper.ResolveStaticCallTarget(pass.TypesInfo, call)
-					if ok && target.Signature.Results().Len() == 1 {
-						addEdge(resultIdx, target, 0)
-					}
-					continue
-				}
-				ident, ok := ast.Unparen(resultExpr).(*ast.Ident)
-				if !ok {
-					continue
-				}
-				v, ok := pass.TypesInfo.ObjectOf(ident).(*types.Var)
-				if !ok || !stableVars[v] {
-					continue
-				}
-				if source, ok := allocationVars[v]; ok {
-					enumerateConcreteReturnEffects(pass, funcObj, resultIdx, source.structType, source.fieldInits, "", effects)
-					continue
-				}
-				if source, ok := resultVars[v]; ok {
-					addEdge(resultIdx, source.callee, source.idx)
-				}
-			}
-			return false
-		}
-		return true
-	})
-}
-
-// enumerateConcreteReturnEffects records omitted and explicitly nil fields from one allocation.
-func enumerateConcreteReturnEffects(pass *analysishelper.EnhancedPass, funcObj *types.Func, resultIdx int, structType *types.Struct, fieldInits []ast.Expr, prefix string, effects fieldEffects) {
-	for i := range structType.NumFields() {
-		field := structType.Field(i)
-		path := joinFieldPath(prefix, field.Name())
-		fieldVal := asthelper.GetFieldVal(fieldInits, field.Name(), structType.NumFields(), i)
-		nilable := !typeshelper.TypeBarsNilness(field.Type())
-		switch {
-		case fieldVal == nil && !nilable:
-			if innerType := typeshelper.AsDeeplyStruct(field.Type()); innerType != nil {
-				enumerateConcreteReturnEffects(pass, funcObj, resultIdx, innerType, nil, path, effects)
-			}
-		case fieldVal == nil || pass.IsNil(fieldVal):
-			if nilable && resultFieldPathIsAcyclic(funcObj, resultIdx, path) {
-				effects.add(funcObj, IndexedFieldPath{Idx: resultIdx, Path: path})
-			}
-		default:
-			if source, ok := staticStructAllocation(pass, fieldVal); ok {
-				enumerateConcreteReturnEffects(pass, funcObj, resultIdx, source.structType, source.fieldInits, path, effects)
-			}
-		}
-	}
-}
-
-// collectFieldReadDemand records the finite field-path demand implied by a single selector
-// expression. There are two distinct demands:
-//
-//   - Base demand: to evaluate `base.Sel`, base must be non-nil, so the field path of base
-//     (relative to the boundary value it roots at) is a read of that value. For `a.y.z` this
-//     records `y`. A selector whose base is the boundary value itself (prefix "") only requires
-//     the value's own top-level nilability, handled by the ordinary annotation machinery.
-//   - Value demand: reading the selected field value itself may later flow through a local
-//     snapshot (`w := a.y; return &D{v: w}`), so for a nilable selected field the full selector
-//     path is recorded. This is what makes `a.y` demand `y`, not only deeper selectors like
-//     `a.y.z`. Suppressed for assignment LHS selectors (skipValueDemand), which write rather
-//     than read the selected field.
-//
-// If the root is a parameter/receiver of funcObj the demand goes to reads (param-in); if it roots
-// at a local bound from a struct-returning call (resultVars) it goes to returnReads, attributed to
-// that callee's result.
-func collectFieldReadDemand(pass *analysishelper.EnhancedPass, sel *ast.SelectorExpr, paramIdx map[*types.Var]int, resultVars map[*types.Var]structResultSource, funcObj *types.Func, reads, returnReads fieldEffects, skipValueDemand bool) {
-	record := func(base *ast.Ident, path string) {
-		if base == nil || path == "" {
-			return
-		}
-		v, ok := pass.TypesInfo.ObjectOf(base).(*types.Var)
-		if !ok {
-			return
-		}
-		if idx, ok := paramIdx[v]; ok {
-			reads.add(funcObj, IndexedFieldPath{Idx: idx, Path: path})
-			return
-		}
-		if src, ok := resultVars[v]; ok {
-			returnReads.add(src.callee.Origin, IndexedFieldPath{Idx: src.idx, Path: path})
-		}
-	}
-
-	// Base demand: `a.y.z` needs `a.y` to be non-nil.
-	base, prefix := asthelper.SplitFieldChain(sel.X)
-	record(base, prefix)
-
-	// Value demand: `a.y` reads field `y` as a value. Only nilable field values need a field
-	// nilability summary; non-nilable value-struct sub-fields are reached by their own later
-	// selectors.
-	field, ok := pass.TypesInfo.ObjectOf(sel.Sel).(*types.Var)
-	if skipValueDemand || !ok || typeshelper.TypeBarsNilness(field.Type()) {
-		return
-	}
-	base, path := asthelper.SplitFieldChain(sel)
-	record(base, path)
-}
-
-// collectParamFieldWrites records nilable fields assigned through a pointer parameter or receiver.
-// The full path is retained for deep and explicit-deref writes. Recursive paths are skipped so the
-// forwarding closure remains finite.
-func collectParamFieldWrites(pass *analysishelper.EnhancedPass, assign *ast.AssignStmt, paramIdx map[*types.Var]int, funcObj *types.Func, writes fieldEffects) {
-	for _, lhs := range assign.Lhs {
-		sel, ok := ast.Unparen(lhs).(*ast.SelectorExpr)
-		if !ok {
-			continue
-		}
-		field, ok := pass.TypesInfo.ObjectOf(sel.Sel).(*types.Var)
-		if !ok || typeshelper.TypeBarsNilness(field.Type()) {
-			continue
-		}
-		base, path := asthelper.SplitFieldChain(lhs)
-		if base == nil || path == "" {
-			continue
-		}
-		param, ok := pass.TypesInfo.ObjectOf(base).(*types.Var)
-		if !ok {
-			continue
-		}
-		idx, ok := paramIdx[param]
-		if !ok {
-			continue
-		}
-		if _, ok := param.Type().Underlying().(*types.Pointer); !ok {
-			continue
-		}
-		if !paramFieldPathIsAcyclic(funcObj, idx, path) {
-			continue
-		}
-		writes.add(funcObj, IndexedFieldPath{Idx: idx, Path: path})
-	}
-}
-
-// collectParamForwardEdges records an arg→param
-// edge for each argument (and the receiver) of call that resolves — through a field chain — to a
-// parameter/receiver of funcObj (the function containing the call). Unresolvable (interface/func-value)
-// callees contribute no edge or callee.
-func collectParamForwardEdges(pass *analysishelper.EnhancedPass, call *ast.CallExpr, paramIdx map[*types.Var]int, funcObj *types.Func, edges map[*types.Func][]paramFieldForwardEdge) *types.Func {
-	target, ok := typeshelper.ResolveStaticCallTarget(pass.TypesInfo, call)
-	if !ok {
-		return nil
-	}
-	record := func(calleeIdx int, arg ast.Expr) {
-		base, prefix := asthelper.SplitFieldChain(arg)
-		if base == nil {
-			return
-		}
-		v, ok := pass.TypesInfo.ObjectOf(base).(*types.Var)
-		if !ok {
-			return
-		}
-		callerIdx, ok := paramIdx[v]
-		if !ok {
-			return
-		}
-		edges[funcObj] = append(edges[funcObj], paramFieldForwardEdge{
-			callerParamIdx: callerIdx,
-			callerPrefix:   prefix,
-			callee:         target.Origin,
-			calleeParamIdx: calleeIdx,
-		})
-	}
-	if recv := target.Signature.Recv(); recv != nil {
-		if sel, ok := ast.Unparen(call.Fun).(*ast.SelectorExpr); ok {
-			record(annotation.ReceiverParamIndex, sel.X)
-		}
-	}
-	for argIdx, arg := range call.Args {
-		if argIdx >= target.Signature.Params().Len() {
-			break
-		}
-		record(argIdx, arg)
-	}
-	return target.Origin
 }
 
 // paramFieldForwardEdge records that the caller passes its own parameter/receiver (callerParamIdx) — possibly
